crypto: move signify to crypto/signify
